+++ core/src/main/java/nl/nn/adapterframework/jms/PushingJmsListener.java
@@ -22,7 +22,10 @@ import javax.jms.Destination;
 import javax.jms.Message;
 import javax.jms.Session;
 
+import org.apache.commons.lang.StringUtils;
+
 import nl.nn.adapterframework.configuration.ConfigurationException;
+import nl.nn.adapterframework.configuration.ConfigurationWarnings;
 import nl.nn.adapterframework.core.IKnowsDeliveryCount;
 import nl.nn.adapterframework.core.IListenerConnector;
 import nl.nn.adapterframework.core.IMessageHandler;
@@ -35,10 +38,6 @@ import nl.nn.adapterframework.core.IbisExceptionListener;
 import nl.nn.adapterframework.core.ListenerException;
 import nl.nn.adapterframework.core.PipeLineResult;
 import nl.nn.adapterframework.receivers.ReceiverBase;
-import nl.nn.adapterframework.unmanaged.SpringJmsConnector;
-
-import org.apache.commons.lang.StringUtils;
-import org.springframework.jms.listener.DefaultMessageListenerContainer;
 
 /**
  * JMSListener re-implemented as a pushing listener rather than a pulling listener.
@@ -70,6 +69,8 @@ import org.springframework.jms.listener.DefaultMessageListenerContainer;
  * <tr><td>{@link #setForceMessageIdAsCorrelationId(boolean) forceMessageIdAsCorrelationId}</td><td>
  * forces that the CorrelationId that is received is ignored and replaced by the messageId that is received. Use this to create a new, globally unique correlationId to be used downstream. It also
  * forces that not the Correlation ID of the received message is used in a reply as CorrelationId, but the MessageId.</td><td>false</td></tr>
+ * <tr><td>{@link #setTimeOut(long) timeOut}</td><td>receive timeout in milliseconds as specified by the JMS API, see https://docs.oracle.com/javaee/7/api/javax/jms/MessageConsumer.html#receive-long-</td><td>1000 [ms]</td></tr>
+ * <tr><td>{@link #setPollGuardInterval(long) pollGuardInterval}</td><td>interval in milliseconds for the poll guard to check whether a successful poll was done by the receive (https://docs.oracle.com/javaee/7/api/javax/jms/MessageConsumer.html#receive-long-) since last check. When polling has stopped this will be logged and the listener will be stopped and started in an attempt to workaround problems with polling. Polling might stop due to bugs in the JMS driver/implementation which should be fixed by the supplier. As the poll time includes reading and processing of the message no successful poll might be registered since the last check when message processing takes a long time, hence while messages are being processed the check on last successful poll will be skipped. Set to -1 to disable</td><td>ten times the specified timeOut</td></tr>
  * </table>
  *</p><p><b>Using transactions</b><br/>
  * This version of the <code>JmsListener</code> supports distributed transactions using the XA-protocol.
@@ -115,19 +116,18 @@ public class PushingJmsListener extends JmsListenerBase implements IPortConnecte
 
 	private String listenerPort;
 	private String cacheMode;
+	private IListenerConnector jmsConnector;
+	private IMessageHandler handler;
+	private IReceiver receiver;
+	private IbisExceptionListener exceptionListener;
+	private long pollGuardInterval = Long.MIN_VALUE;
 
-    private IListenerConnector jmsConnector;
-    private IMessageHandler handler;
-    private IReceiver receiver;
-    private IbisExceptionListener exceptionListener;
-
-
-    @Override
-    public void configure() throws ConfigurationException {
-        super.configure();
-        if (jmsConnector==null) {
-        	throw new ConfigurationException(getLogPrefix()+" has no jmsConnector. It should be configured via springContext.xml");
-        }
+	@Override
+	public void configure() throws ConfigurationException {
+		super.configure();
+		if (jmsConnector==null) {
+			throw new ConfigurationException(getLogPrefix()+" has no jmsConnector. It should be configured via springContext.xml");
+		}
 		if (StringUtils.isNotEmpty(getCacheMode())) {
 			if (!getCacheMode().equals("CACHE_NONE") &&
 				!getCacheMode().equals("CACHE_CONNECTION") &&
@@ -137,17 +137,27 @@ public class PushingJmsListener extends JmsListenerBase implements IPortConnecte
 				}
 		}
 		Destination destination;
-        try {
+		try {
 			destination = getDestination();
 		} catch (Exception e) {
 			throw new ConfigurationException(getLogPrefix()+"could not get Destination",e);
 		}
-        try {
-			jmsConnector.configureEndpointConnection(this, getMessagingSource().getConnectionFactory(), destination, getExceptionListener(), getCacheMode(), getAckMode(), isJmsTransacted(), getMessageSelector());
+		if (getPollGuardInterval() == Long.MIN_VALUE) {
+			setPollGuardInterval(getTimeOut() * 10);
+		}
+		if (getPollGuardInterval() <= getTimeOut()) {
+			ConfigurationWarnings configWarnings = ConfigurationWarnings.getInstance();
+			configWarnings.add(log, "The pollGuardInterval [" + getPollGuardInterval()
+					+ "] should be larger than the receive timeout [" + getTimeOut() + "]");
+		}
+		try {
+			jmsConnector.configureEndpointConnection(this, getMessagingSource().getConnectionFactory(), destination,
+					getExceptionListener(), getCacheMode(), getAckMode(), isJmsTransacted(),
+					getMessageSelector(), getTimeOut(), getPollGuardInterval());
 		} catch (JmsException e) {
 			throw new ConfigurationException(e);
 		}
-    }
+	}
 
     @Override
     public void open() throws ListenerException {
@@ -166,25 +176,6 @@ public class PushingJmsListener extends JmsListenerBase implements IPortConnecte
 		}
 	}
 
-    public void destroy() {
-    	if (jmsConnector!=null) {
-    		if (jmsConnector instanceof SpringJmsConnector) {
-    			SpringJmsConnector springJmsConnector = (SpringJmsConnector) jmsConnector;
-    			DefaultMessageListenerContainer jmsContainer = springJmsConnector.getJmsContainer();
-    	    	if (jmsContainer!=null) {
-        			jmsContainer.destroy();
-    	    		log.debug(getLogPrefix()+"jmsContainer is destroyed");
-    	    	} else {
-    	    		log.error(getLogPrefix()+"could not destroy, jmsContainer is null");
-    	    	}
-        	} else {
-        		log.error(getLogPrefix()+"could not destroy, jmsConnector is not instance of SpringJmsConnector but ["+jmsConnector.getClass().getName()+"]");
-    		}
-    	} else {
-    		log.error(getLogPrefix()+"could not destroy, jmsConnector is null");
-    	}
-    }
-    
 	public void afterMessageProcessed(PipeLineResult plr, Object rawMessage, Map threadContext) throws ListenerException {
 		String cid     = (String) threadContext.get(IPipeLineSession.technicalCorrelationIdKey);
 		Session session= (Session) threadContext.get(jmsConnector.THREAD_CONTEXT_SESSION_KEY); // session is/must be saved in threadcontext by JmsConnector
@@ -395,6 +386,14 @@ public class PushingJmsListener extends JmsListenerBase implements IPortConnecte
 			log.error(getLogPrefix()+"exception in determination of DeliveryCount", e);
 			return -1;
 		}
+	}
+
+	public void setPollGuardInterval(long pollGuardInterval) {
+		this.pollGuardInterval = pollGuardInterval;
+	}
+
+	public long getPollGuardInterval() {
+		return pollGuardInterval;
 	}
 
 }